requirements.txt
proxmoxer
getpass4
console-menu
pyyaml